Add GitHub CI workflow (#15)
This commit adds a GitHub CI workflow which configures and compiles the code on an Ubuntu runner with CMake and an Intel compiler and MPI implementation. After compilation, a basic channel flow case is ran for a short time. This PR also changes the MPI interfacing from the outdated include mpif.h to the more recommended use mpi (a few more changes are needed to move to use mpi_f08).

Based on the squashed commits:

* working CI

* added short parallel run of the executable

* added mkdir channel_output to run example

* added shell:bash to check if this ensures that -o pipefail is applied and mpi errors cause exiting

* check if || exit 1 helps

* added print statement for debugging

* change CI MPI installation

* fixed formatting

* check if different results are obtained when changing to use mpi

* switch to intel mpi

* try building fftw from source instead

* removed FFTW mpi dependeny installation

* Last commit works. Try caching FFTW to speed up

* Previous commit takes too long, try installing fftw in /opt/fftw3

* added echo FC=ifx >>