Change body_type from integer to string (#29)
This PR changes the type of the body_type variable from an Integer(Int32) to a Character(200). The corresponding input is changed to lower case immediately after the input is read. Every subsequent time that body_type is used, it appears as trim(body_type) to compare it to different values without trailing white space.